fixing refactoring of HasVisibleHandlers and changed the construction of CollapseTrigger when inside Accordion
